+++ data-access/nexustiles/AbstractTileService.py
@@ -19,24 +19,31 @@ from functools import reduce
 import numpy as np
 import numpy.ma as ma
 
+from datetime import datetime
+
 
 class AbstractTileService(ABC):
     def __init__(self, dataset_name):
         self._name = dataset_name
+        self._last_updated = datetime.now()
 
     def heartbeat(self) -> bool:
         return True
 
-    def update(self, force: bool=False) -> bool:
+    def update(self, force: bool = False, load: bool = False) -> bool:
         """
         If applicable, verify the underlying connection(s) are still open and valid, replacing them if necessary
 
         Default implementation is to do nothing and assume all is valid.
 
+        @param load:
         @param force: Forcibly replace underlying connections.
         @return: True if valid and successful, False if dataset is now unreachable and therefore should not be used
         """
         return True
+
+    def last_updated(self):
+        return self._last_updated
 
     @abstractmethod
     def get_dataseries_list(self, simple=False):